utils.py
import os
import logging
from dotenv import load_dotenv

load_dotenv()

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s %(levelname)s %(name)s - %(message)s"
)

def env(name: str, default: str = "") -> str:
    return os.getenv(name, default)
