requirements.txt
aiogram>=3.4.1
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
selectolax>=0.3.22
pdfplumber>=0.11.4
camelot-py[cv]>=0.11.0
tabula-py>=2.10.0
pandas>=2.2.2
openpyxl>=3.1.4
pydantic>=2.7.4
chromadb>=0.5.3
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
scikit-learn>=1.5.1
structlog>=24.1.0
pytest>=8.2.1
